requirements.txt
flask
Flask-Script
Flask-SQLAlchemy
Flask-Migrate
Flask-Testing
